requirements.txt
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
channels==4.3.2
channels_redis==4.3.0
click==8.1.8
dj-database-url==3.1.2
Django==5.0.10
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==24.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
gunicorn==25.3.0
h11==0.16.0
httptools==0.7.1
idna==3.13
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==26.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.45.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
whitenoise==6.12.0
